+++ Hockey Playoffs/Bracket/BracketLine.swift
@@ -14,7 +14,7 @@ class BracketLine: UIView {
     }
     
     init() {
-        super.init(frame: CGRect.zero)
+        super.init(frame: .zero)
         translatesAutoresizingMaskIntoConstraints = false
         
         let view = lineView()
@@ -28,7 +28,7 @@ class BracketLine: UIView {
     }
     
     init(orientation: BracketOrientation) {
-        super.init(frame: CGRect.zero)
+        super.init(frame: .zero)
         translatesAutoresizingMaskIntoConstraints = false
         
         let topLeading = lineView()
